requirements.txt
aiodns==2.0.0
aiofiles==0.5.0
aiohttp==3.6.2
aioinflux==0.9.0
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
ciso8601==2.1.3
idna==2.9
multidict==4.7.6
pycares==3.1.1
pycparser==2.20
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
yarl==1.4.2
